hrb-lbo/requirements.txt
pandas>=1.5.0
numpy>=1.20.0
pyyaml>=6.0
streamlit>=1.28.0
plotly>=5.18.0
